#11: Move layout to Main.tsx
